requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
rplidar==0.9.2
rplidar-roboticia==0.9.5
six==1.16.0
